pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "auto-gpt-obsidian-plugin"
version = "0.0.2"
description = "A plugin for autogpt that allows interaction with Obsidian."
authors = ["Conner Ohnesorge <conneroisu@outlook.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/conneroisu/Auto-GPT-Obsidian"
repository = "https://github.com/conneroisu/Auto-GPT-Obsidian"
documentation = "https://github.com/conneroisu/Auto-GPT-Obsidian"
keywords = ["autogpt", "obsidian", "plugin"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
packages = [{ include = "obsidian", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.8"
obsidiantools = "*"
python-dotenv = "^1.0.1"
auto_gpt_plugin_template = "*"
pydantic = "^2.7.1"
llama-cpp-python = "^0.2.79"
chromadb = "^0.5.3"
pyyaml = "^6.0.1"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
gitignore-parser = "^0.1.11"
jinja2 = "^3.1.4"
unidecode = "^1.3.8"
diff-match-patch = "^20230430"

[tool.poetry.group.dev.dependencies]
pytest = "*"
flake8 = "*"
black = "*"
isort = "*"
pytest-mock = "^3.12.0"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\\.pyi?$'

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 88
extend-ignore = "E203, W503"
